fix(module): ensure var names are well-formed
by adding a `shellVarType` type that checks `name` and `namePath`
against the pattern that conforming shell variable names must satisfy.

Additionally, this change introduces coercion the default `name` and
`namePath` values to valid shell variable names by (a) replacing invalid
characters with `__` and (b) prepending `__` in case the value would
otherwise start with an (illegal) digit.

This should not be a breaking change, as shellcheck already rejects
nonconforming shell variable names.  This change merely promotes that
check to the NixOS option type level, and provides a convenience feature
for entries in `agenix-shell.secrets` whose (attribute) names may not be
valid shell variable names.